tema07_docker/08_compose/server/requirements.txt
Flask==2.0.2
Flask-CLI==0.4.0
mysql-connector-python==8.0.27
